+++ backend/src/main/java/kr/momo/config/interceptor/LogInterceptor.java
@@ -1,4 +1,4 @@
-package kr.momo.config.filter;
+package kr.momo.config.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
